exercise 3 done

+++ week-06/day-1/src/Exercuse2.java
@@ -8,7 +8,7 @@ public class Exercuse2 {
     ArrayList<Integer> numbers = new ArrayList<>(Arrays.asList(1, 3, -2, -4, -7, -3, -8, 12, 19, 6, 9, 10, 14));
     System.out.println(numbers.stream()
             .filter(n -> n % 2 != 0)
-            .mapToInt(n -> n)
+            .mapToDouble(n -> n)
             .average());
   }
 }
